Migrated project to Vite from Create-React-App
Migrated from CRA to Vite in order to create faster build times and to avoid issues with webpack moving forward. Eventually will update this project in order to handle for a merchandising product table that will allow for Users to purchase merchandise, search for it, and browse items related to specific categories.